remove all drake utilities and functions and replace with gurobipy - also applied formatting
